backend/requirements.txt
praw==7.7.1
python-dotenv==1.0.0
ollama==0.1.8
fastapi
uvicorn
groq
python-cors
